+++ src/OpenLoco/src/Input.cpp
@@ -15,6 +15,7 @@ namespace OpenLoco::Input
     static Ui::Point _cursorDragStart;
     static uint32_t _cursorDragState;
     static bool _exitRequested = false;
+    static bool _isLostFocusPaused = false;
 
     void init()
     {
@@ -45,6 +46,33 @@ namespace OpenLoco::Input
     void state(State state)
     {
         _state = state;
+    }
+
+    static void focusLost()
+    {
+        if (!Config::get().pauseOnLostFocus)
+        {
+            return;
+        }
+
+        _isLostFocusPaused = true;
+        Audio::toggleAudioLostFocus(true);
+    }
+
+    static void focusGained()
+    {
+        if (!_isLostFocusPaused)
+        {
+            return;
+        }
+
+        _isLostFocusPaused = false;
+        Audio::toggleAudioLostFocus(false);
+    }
+
+    bool isLostFocusPaused()
+    {
+        return _isLostFocusPaused;
     }
 
     // 0x00407218
@@ -129,6 +157,12 @@ namespace OpenLoco::Input
 
                 case SDL_EVENT_WINDOW_MOVED:
                     Ui::windowPositionChanged(e.window.data1, e.window.data2);
+                    break;
+                case SDL_WINDOWEVENT_FOCUS_LOST:
+                    focusLost();
+                    break;
+                case SDL_WINDOWEVENT_FOCUS_GAINED:
+                    focusGained();
                     break;
                 case SDL_EVENT_WINDOW_RESIZED:
                 case SDL_EVENT_WINDOW_PIXEL_SIZE_CHANGED: