+++ src/OpenLoco/src/Localisation/StringIds.h
@@ -2150,6 +2150,8 @@ namespace OpenLoco::StringIds
     constexpr StringId cheat_vehicle_cargo = 2434;
     constexpr StringId cheat_keep_cargo_modify_pickup = 2435;
     constexpr StringId tooltip_keep_cargo_modify_pickup = 2436;
+    constexpr StringId pause_on_lost_focus = 2437;
+    constexpr StringId pause_on_lost_focus_tip = 2438;
 
     constexpr StringId temporary_object_load_str_0 = 8192;
     constexpr StringId temporary_object_load_str_1 = 8193;